Logout, Sign up, Categories, Status
I made Log in and Sign up but it is not done yet (not working as it should)
I fixed the categories and you can filter the tasks by category. you can set a category from a list when you create task.
I fixed the status and you can filter the tasks by status as well change the status of the task with a popin modal where you can choose a status and send the task to the column.
The functions are not in the final state. Needs inpruvment.